Ability to add app specific exclusions
We had an app with a very short name that triggered every week as matching vulnerabilities (the word matched parts of other words). This update allows the admin to add additional lines to the INI file (manually) that will exclude an application from vulnerability checks. Obviously we don't want to exclude anything but if needed, it is possible now.