Bug Fix
It looks like MySQL Server 8.0.16 and above does not support querying on blank or empty date values (NULL is ok). Not sure why we where looking for them but I'm sure there was a reason originally. Since I can't think of a good reason to keep querying for them, that has been removed from the query.