Fix Version_Newest not installed
An attempt to fix when the newest discovered version is not installed. This happens when the length of the version number changes or if the newest version is later uninstalled. Not sure if we want to remove the latter of those. But if the length is different it causes issues in a few other places. At this point, I think this still needs work.