Software update report links added
Added links to fix possible updates listed that need verification. If there is no update available and the web GUI option is installed/selected, the user can click a link and confirm that it was reviewed and the database will be updated with the new location of the version on the web page.