AppRename Table Added
Added the AppRename table. For now, it can be manually edited to supply regex to match application names and rename them so they are all named the same. I would use this sparingly but it comes in handy if, for instance, you have an application that appends the PC name to the application name in list of installed programs. Rather than ending up with hundreds of unique apps, you can have them all show up as the same app.