Hide removed software
Changed the code to hide the software that have not been seen as of the latest pull. If something was removed within the last 7 days it remains in the database to help with inventorying new versions or in case it is reinstalled. I've decided we don't really want those applications to show up on the update report email though.